Conflicts:
	build/cocos2d_libs.xcodeproj/project.pbxproj
	cocos/scripting/lua/bindings/CCLuaStack.cpp
	cocos/scripting/lua/script/Cocos2d.lua
	cocos/scripting/lua/script/Cocos2dConstants.lua

+++ cocos/2d/CCTextFieldTTF.cpp
@@ -216,7 +216,7 @@ void TextFieldTTF::deleteBackward()
         ++deleteLen;
     }
 
-    if (_delegate && _delegate->onTextFieldDeleteBackward(this, _inputText.c_str() + len - deleteLen, deleteLen))
+    if (_delegate && _delegate->onTextFieldDeleteBackward(this, _inputText.c_str() + len - deleteLen, static_cast<int>(deleteLen)))
     {
         // delegate doesn't wan't to delete backwards
         return;
